+++ flint/ms.py
@@ -504,6 +504,55 @@ def consistent_ms(ms1: MS, ms2: MS) -> bool:
     return result
 
 
+def consistent_channelwise_frequencies(
+    freqs: Union[List[np.ndarray], np.ndarray],
+) -> np.ndarray:
+    """Given a collection of frequencies in the form of
+    (N, frequencies), inspect the frequencies channelwise
+    to ensure they are all the same.
+
+    This does not operate on MSs, just the collection of frequencies
+
+    Args:
+        freqs (Union[List[np.ndarray], np.ndarray]): The collection of frequencies to be inspected
+
+    Returns:
+        np.ndarray: Same length as the frequencies. True if for a single channel all frequencies are the same. False otherwise.
+    """
+    freqs = np.array(freqs)
+    assert (
+        len(freqs.shape) == 2
+    ), f"{freqs.shape=}, but was expecting something of rank 2"
+
+    freqs_are_same = np.all(freqs - freqs[0, None] == 0, axis=1)
+    assert (
+        len(freqs_are_same.shape) == 1
+    ), f"Channelwise check should be length 1, but have {freqs_are_same.shaope=}"
+    return freqs_are_same
+
+
+def consistent_ms_frequencies(mss: Tuple[MS, ...]) -> bool:
+    """Given a set of measurement sets, inspect the frequencies
+    to ensure they are all the same
+
+    See the ``get_freqs_from_ms`` function, which is used
+    internally.
+
+    Args:
+        mss (Tuple[MS, ...]): Collection of MSs to inspect the frequencies of
+
+    Returns:
+        bool: Whether all the frequencies and their order are the same
+    """
+
+    logger.info(f"Collection frequencies from {len(mss)} measurement sets")
+    freqs = [get_freqs_from_ms(ms=ms) for ms in mss]
+
+    all_the_same = consistent_channelwise_frequencies(freqs=freqs)
+
+    return np.all(all_the_same)
+
+
 def rename_column_in_ms(
     ms: MS,
     original_column_name: str,
@@ -574,7 +623,11 @@ def remove_columns_from_ms(
 
 
 def subtract_model_from_data_column(
-    ms: MS, model_column: str = "MODEL_DATA", data_column: Optional[str] = None
+    ms: MS,
+    model_column: str = "MODEL_DATA",
+    data_column: Optional[str] = None,
+    output_column: Optional[str] = None,
+    update_tracked_column: bool = False,
 ) -> MS:
     """Execute a ``taql`` query to subtract the MODEL_DATA from a nominated data column.
     This requires the ``model_column`` to already be inserted into the MS. Internally
@@ -585,6 +638,8 @@ def subtract_model_from_data_column(
         ms (MS): The measurement set instance being considered
         model_column (str, optional): The column with representing the model. Defaults to "MODEL_DATA".
         data_column (Optional[str], optional): The column where the column will be subtracted. If ``None`` it is taken from the ``column`` nominated by the input ``MS`` instance. Defaults to None.
+        output_column (Optional[str], optional): The output column that will be created. If ``None`` it defaults to ``data_column``. Defaults to None.
+        update_tracked_column (bool, optional): If True, update ``ms.column`` to the column with subtracted data. Defaults to False.
 
     Returns:
         MS: The updated MS
@@ -592,6 +647,9 @@ def subtract_model_from_data_column(
     ms = MS.cast(ms)
     data_column = data_column if data_column else ms.column
     assert data_column is not None, f"{data_column=}, which is not allowed"
+
+    output_column = output_column if output_column else data_column
+    assert output_column is not None, f"{output_column=}, which is not allowed"
     with critical_ms_interaction(input_ms=ms.path) as critical_ms:
         with table(str(critical_ms), readonly=False) as tab:
             logger.info("Extracting columns")
@@ -600,12 +658,27 @@ def subtract_model_from_data_column(
                 [d in colnames for d in (model_column, data_column)]
             ), f"{model_column=} or {data_column=} missing from {colnames=}"
 
-            logger.info(f"Subtracting {model_column=} from {data_column=}")
-            taql(f"UPDATE $tab SET {data_column}={data_column}-{model_column}")
+            if output_column not in colnames:
+                from casacore.tables import makecoldesc
 
+                logger.info(f"Adding {output_column=}")
+                desc = makecoldesc(data_column, tab.getcoldesc(data_column))
+                desc["name"] = output_column
+                tab.addcols(desc)
+                tab.flush()
+
+            logger.info(f"Subtracting {model_column=} from {data_column=}")
+            taql(f"UPDATE $tab SET {output_column}={data_column}-{model_column}")
+
+    if update_tracked_column:
+        logger.info(f"Updating ms.column to {output_column=}")
+        ms = ms.with_options(column=output_column)
     return ms
 
 
+# TODO: Clean up the usage and description of the argument `instrument_column`
+# as it is currently being used in unclear ways. Specifically there is a renaming
+# of the data_column to instrument_column before the rotation of things
 def preprocess_askap_ms(
     ms: Union[MS, Path],
     data_column: str = "DATA",
@@ -848,7 +921,10 @@ def rename_ms_and_columns_for_selfcal(
 
 
 def find_mss(
-    mss_parent_path: Path, expected_ms_count: Optional[int] = 36
+    mss_parent_path: Path,
+    expected_ms_count: Optional[int] = 36,
+    data_column: Optional[str] = None,
+    model_column: Optional[str] = None,
 ) -> Tuple[MS, ...]:
     """Search a directory to find measurement sets via a simple
     `*.ms` glob expression. An expected number of MSs can be enforced
@@ -857,6 +933,8 @@ def find_mss(
     Args:
         mss_parent_path (Path): The parent directory that will be globbed to search for MSs.
         expected_ms_count (Optional[int], optional): The number of MSs that should be there. If None no check is performed. Defaults to 36.
+        data_column (Optional[str], optional): Set the column attribute of each MS to this (no checks to ensure it exists). If None use default of MS. Defaults to None.
+        model_column (Optional[str], optional): Set the model column attribute of each MS to this (no checks to ensure it exists). If None use default of MS. Defaults to None.
 
     Returns:
         Tuple[MS, ...]: Collection of found MSs
@@ -873,6 +951,15 @@ def find_mss(
         assert (
             len(found_mss) == expected_ms_count
         ), f"Expected to find {expected_ms_count} in {str(mss_parent_path)}, found {len(found_mss)}."
+
+    if data_column or model_column:
+        logger.info(f"Updating column attribute to {data_column=}")
+        found_mss = tuple(
+            [
+                found_ms.with_options(column=data_column, model_column=model_column)
+                for found_ms in found_mss
+            ]
+        )
 
     return found_mss